+++ src/euphorie/content/tests/test_country.py
@@ -18,3 +18,10 @@ class CountryTests(EuphorieIntegrationTestCase):
         self.assertEqual(
             country.risk_default_collapsible_sections, ["collapsible_section_information"]
         )
+
+    def testDefaultReports(self):
+        country = self.createCountry()
+        self.assertEqual(
+            country.default_reports,
+            ["report_full", "report_action_plan", "report_overview_risks"],
+        )